notification and alert

+++ app/src/main/java/com/example/weatherapp/model/pojo/WeatherDto.kt
@@ -12,5 +12,5 @@ data class WeatherDto(
     val lat: Double,
     val lon: Double,
     val timezone: String,
-    val timezone_offset: Int
+    val timezone_offset: Double
 )
